--- app/controllers/admin_controller.rb
@@ -1,13 +1,0 @@
-class AdminController < ApplicationController
-  before_action :authenticate_user!
-  before_action :require_admin
-
-  protected
-
-  def require_admin
-    return if current_user.admin?
-
-    flash[:alert] = t('forbidden')
-    redirect_to root_path
-  end
-end
