+++ app/controllers/application_controller.rb
@@ -1,5 +1,6 @@
 class ApplicationController < ActionController::Base
   before_action :configure_permited_parameters, if: :devise_controller?
+  before_action :authenticate_user!
 
   protected
 
@@ -7,9 +8,17 @@ class ApplicationController < ActionController::Base
     devise_parameter_sanitizer.permit(:sign_up, keys: [:cpf])
   end
 
-  def authorize_admin
-    return unless current_user.nil? || !current_user.admin?
+  def require_manager
+    return if current_user.manager?
 
-    redirect_to root_path, alert: t('errors.user.permission_denied')
+    flash[:alert] = t('forbidden')
+    redirect_to root_path
+  end
+
+  def require_admin
+    return if current_user.admin?
+
+    flash[:alert] = t('forbidden')
+    redirect_to root_path
   end
 end
